Revert "Serum: Integrating serum_gov & DAO-to-DAO interaction (#961)"
This reverts commit 33bc0c5a6b1951eb60b36e36f061bfa862b7afb6.
