+++ components/TokenBalance/TokenBalanceCard.tsx
@@ -37,7 +37,6 @@ import useNftPluginStore from 'NftVotePlugin/store/nftPluginStore'
 import { vsrPluginsPks } from '@hooks/useVotingPlugins'
 import { REALM_ID as PYTH_REALM_ID, PythBalance } from 'pyth-staking-api'
 import DelegateTokenBalanceCard from '@components/TokenBalance/DelegateTokenBalanceCard'
-import SerumGovernanceTokenWrapper from './SerumGovernanceTokenWrapper'
 import getNumTokens from '@components/ProposalVotingPower/getNumTokens'
 import VotingPowerPct from '@components/ProposalVotingPower/VotingPowerPct'
 
@@ -50,7 +49,6 @@ const TokenBalanceCard = ({
   inAccountDetails?: boolean
   children?: React.ReactNode
 }) => {
-  const realmProgramId = useWalletStore((s) => s.selectedRealm.programId)
   const [hasGovPower, setHasGovPower] = useState<boolean>(false)
   const { councilMint, mint, realm } = useRealm()
   const connected = useWalletStore((s) => s.connected)
@@ -118,11 +116,6 @@ const TokenBalanceCard = ({
           <div className="h-10 rounded-lg animate-pulse bg-bkg-3" />
         </>
       )}
-      {/* TODO: Restrict to Serum DAO */}
-      {realmProgramId?.toBase58() ===
-      'G41fmJzd29v7Qmdi8ZyTBBYa98ghh3cwHBTexqCG1PQJ' ? (
-        <SerumGovernanceTokenWrapper />
-      ) : null}
       {children}
     </>
   )
