+++ components/VoteResults.tsx
@@ -1,6 +1,6 @@
 import VoteResultsBar from './VoteResultsBar'
-import { Proposal } from '@solana/spl-governance'
 import useProposalVotes from '@hooks/useProposalVotes'
+import { Proposal } from '@solana/spl-governance'
 
 type VoteResultsProps = {
   isListView?: boolean
@@ -14,7 +14,6 @@ const VoteResults = ({ isListView, proposal }: VoteResultsProps) => {
     relativeNoVotes,
     relativeYesVotes,
   } = useProposalVotes(proposal)
-
   return (
     <div className="flex items-center space-x-4">
       {proposal ? (
