+++ components/instructions/programs/names.ts
@@ -8,12 +8,7 @@ import { NAME_PROGRAM_ID } from '@bonfida/spl-name-service'
 
 export const GOVERNANCE_PROGRAM_NAMES = {
   GqTPL6qRf5aUuqscLh8Rg2HTxPUXfhhAXDptTLhp1t2J: 'Mango Governance Program',
-  AVoAYTs36yB5izAaBkxRG67wL1AMwG3vo41hKtUSb8is:
-    'Serum Governance Program (Old)',
-  G41fmJzd29v7Qmdi8ZyTBBYa98ghh3cwHBTexqCG1PQJ:
-    'Serum Governance Program (New)',
-  FBcTbv5rLy7MQkkAU2uDzAEjjZDeu2BVLVRJGxyz6hnV:
-    'Serum Governance Token Program',
+  AVoAYTs36yB5izAaBkxRG67wL1AMwG3vo41hKtUSb8is: 'Serum Governance Program',
   GTesTBiEWE32WHXXE2S4XbZvA5CrEc4xs6ZgRe895dP: 'Test Governance Program',
   GovER5Lthms3bLBqWub97yVrMmEogzX7xNjdXpPPCVZw: 'Governance Program',
   '5sGZEdn32y8nHax7TxEyoHuPS3UXfPWtisgm8kqxat8H':
