+++ components/treasuryV2/Details/index.tsx
@@ -13,7 +13,6 @@ import ProgramsDetails from './ProgramsDetails'
 import RealmAuthorityDetails from './RealmAuthorityDetails'
 import TokenDetails from './TokenDetails'
 import WalletDetails from './WalletDetails'
-import TokenOwnerRecordDetails from './TokenOwnerRecordDetails'
 
 function walletIsNotAuxiliary(
   wallet: AuxiliaryWallet | Wallet
@@ -90,14 +89,6 @@ const Details = forwardRef<HTMLDivElement, Props>((props, ref) => {
                 <NFTCollectionDetails
                   nftCollection={props.data.data.asset}
                   isStickied={props.isStickied}
-                />
-              ) : props.data.data.asset.type ===
-                  AssetType.TokenOwnerRecordAsset &&
-                walletIsNotAuxiliary(props.data.data.wallet) ? (
-                <TokenOwnerRecordDetails
-                  isStickied={props.isStickied}
-                  selectedWallet={props.data.data.wallet}
-                  tokenOwnerRecordAsset={props.data.data.asset}
                 />
               ) : (
                 <div />
