+++ components/treasuryV2/WalletList/WalletListItem/AssetList/index.tsx
@@ -24,14 +24,12 @@ import {
   isPrograms,
   isRealmAuthority,
   isUnknown,
-  isTokenOwnerRecord,
 } from '../typeGuards'
 
 import { PublicKey } from '@solana/web3.js'
 import { Metadata } from '@metaplex-foundation/mpl-token-metadata'
 import { findMetadataPda } from '@metaplex-foundation/js'
 import useWalletStore from 'stores/useWalletStore'
-import TokenOwnerRecordsList from './TokenOwnerRecordsList'
 
 export type Section = 'tokens' | 'nfts' | 'others'
 
@@ -137,11 +135,6 @@ export default function AssetList(props: Props) {
 
   const othersFromProps = useMemo(() => props.assets.filter(isOther), [])
 
-  const tokenOwnerRecordsFromProps = useMemo(
-    () => props.assets.filter(isTokenOwnerRecord),
-    [props.assets]
-  )
-
   const [others, setOthers] = useState<
     (Mint | Programs | Unknown | RealmAuthority)[]
   >(othersFromProps)
@@ -235,16 +228,6 @@ export default function AssetList(props: Props) {
           disableCollapse={!diplayingMultipleAssetTypes}
           expanded={props.expandedSections?.includes('others')}
           assets={others}
-          selectedAssetId={props.selectedAssetId}
-          onSelect={props.onSelectAsset}
-          onToggleExpand={() => props.onToggleExpandSection?.('others')}
-        />
-      )}
-      {tokenOwnerRecordsFromProps.length > 0 && (
-        <TokenOwnerRecordsList
-          disableCollapse={false}
-          expanded={true}
-          assets={tokenOwnerRecordsFromProps}
           selectedAssetId={props.selectedAssetId}
           onSelect={props.onSelectAsset}
           onToggleExpand={() => props.onToggleExpandSection?.('others')}
