+++ components/treasuryV2/WalletList/WalletListItem/index.tsx
@@ -6,7 +6,6 @@ import { Wallet } from '@models/treasury/Wallet'
 
 import AssetList, { Section } from './AssetList'
 import SummaryButton from './SummaryButton'
-import SerumGovWallet from './SerumGovWallet'
 
 interface Props {
   className?: string
@@ -86,7 +85,6 @@ export default function WalletListItem(props: Props) {
               })
             }
           />
-          <SerumGovWallet wallet={props.wallet} />
         </div>
       )}
     </div>
