+++ components/treasuryV2/WalletList/WalletListItem/typeGuards.ts
@@ -8,7 +8,6 @@ import {
   RealmAuthority,
   Sol,
   Unknown,
-  TokenOwnerRecordAsset,
 } from '@models/treasury/Asset'
 
 export function isToken(asset: Asset): asset is Token {
@@ -37,10 +36,4 @@ export function isSol(asset: Asset): asset is Sol {
 
 export function isUnknown(asset: Asset): asset is Unknown {
   return asset.type === AssetType.Unknown
-}
-
-export function isTokenOwnerRecord(
-  asset: Asset
-): asset is TokenOwnerRecordAsset {
-  return asset.type === AssetType.TokenOwnerRecordAsset
 }
