+++ hooks/useTreasuryInfo/index.tsx
@@ -26,7 +26,7 @@ interface Data {
 
 export default function useTreasuryInfo(): Result<Data> {
   const { realmInfo, realm, mint, councilMint, config } = useRealm()
-  const connection = useWalletStore((s) => s.connection)
+  const connection = useWalletStore((s) => s.connection.current)
   const accounts = useGovernanceAssetsStore((s) => s.assetAccounts)
   const loadingGovernedAccounts = useGovernanceAssetsStore(
     (s) => s.loadGovernedAccounts
@@ -83,7 +83,7 @@ export default function useTreasuryInfo(): Result<Data> {
         realmInfo
       )
     }
-  }, [accounts, nfts, nftsLoading, realmInfo, connection])
+  }, [accounts, nfts, nftsLoading, realmInfo])
 
   useEffect(() => {
     setBuildingWallets(true)
