+++ models/treasury/Asset.ts
@@ -1,18 +1,10 @@
 import type { BigNumber } from 'bignumber.js'
-import type {
-  Governance,
-  MintMaxVoteWeightSource,
-  ProgramAccount,
-  Realm,
-  TokenOwnerRecord,
-} from '@solana/spl-governance'
+import type { MintMaxVoteWeightSource } from '@solana/spl-governance'
 
 import type { AssetAccount } from '@utils/uiTypes/assets'
 
 import { NFT } from './NFT'
 import { Program } from './Program'
-import { TokenProgramAccount } from '@utils/tokens'
-import { MintInfo } from '@solana/spl-token'
 import { PublicKey } from '@solana/web3.js'
 
 export enum AssetType {
@@ -23,7 +15,6 @@ export enum AssetType {
   Sol,
   Token,
   Unknown,
-  TokenOwnerRecordAsset,
 }
 
 export interface Mint {
@@ -104,22 +95,6 @@ export interface Unknown {
   name: string
 }
 
-export interface TokenOwnerRecordAsset {
-  type: AssetType.TokenOwnerRecordAsset
-  id: string
-  address: PublicKey
-  owner: PublicKey
-  realmId: string
-  realmSymbol: string
-  displayName: string
-  programId: string
-  realmImage?: string
-  communityMint: TokenProgramAccount<MintInfo>
-  realmAccount: ProgramAccount<Realm>
-  tokenOwnerRecordAccount: ProgramAccount<TokenOwnerRecord>
-  governanceOwner: ProgramAccount<Governance>
-}
-
 export type Asset =
   | Mint
   | NFTCollection
@@ -128,4 +103,3 @@ export type Asset =
   | Sol
   | Token
   | Unknown
-  | TokenOwnerRecordAsset
