+++ pages/_app.tsx
@@ -7,19 +7,8 @@ import { App as HubApp } from '@hub/App'
 import '../styles/index.css'
 import '../styles/typography.css'
 import '@hub/components/controls/RichTextEditor/index.css'
-import { useEffect } from 'react'
-import useSerumGovStore from 'stores/useSerumGovStore'
 
 export default function App({ Component, pageProps, router }: AppProps) {
-  const { cluster } = router.query
-  const updateSerumGovAccounts = useSerumGovStore(
-    (s) => s.actions.updateSerumGovAccounts
-  )
-
-  useEffect(() => {
-    updateSerumGovAccounts(cluster as string | undefined)
-  }, [cluster])
-
   if (
     router.pathname.startsWith('/realm/[id]') ||
     router.pathname.startsWith('/ecosystem') ||
