+++ pages/realms/components/RealmsGrid.tsx
@@ -238,7 +238,6 @@ function RealmsGrid({
                   ) && (
                     <div
                       draggable={editing}
-                      // eslint-disable-next-line react/no-unknown-property
                       unselectable="on"
                       onMouseDown={
                         editing
