+++ public/realms/devnet.json
@@ -1,5 +1,4 @@
 [
-   
   {
     "symbol": "UNQ",
     "displayName": "UNQ Universe",
@@ -17,20 +16,6 @@
     "twitter": "@mangomarkets",
     "ogImage": "https://trade.mango.markets/assets/icons/logo.svg",
     "notificationsEnabled": true
-  },
-  {
-    "symbol": "SERUM",
-    "displayName": "SerumDAO - Devnet",
-    "programId": "G41fmJzd29v7Qmdi8ZyTBBYa98ghh3cwHBTexqCG1PQJ",
-    "realmId": "8rTohXsU2L3xaf3KeHiS55DNninzCjDZT8zkpwikHSQt",
-    "ogImage": "https://assets.website-files.com/61284dcff241c2f0729af9f3/61285237ce2e301255d09108_logo-serum.png"
-  },
-  {
-    "symbol": "DAO2DAO",
-    "displayName": "DAO-to-DAO",
-    "programId": "GTesTBiEWE32WHXXE2S4XbZvA5CrEc4xs6ZgRe895dP",
-    "realmId": "3SXQuce59HJMy1udz9UaQJKjWrRw2s6q2rEK1S9Ep7cv",
-    "ogImage": "https://assets.website-files.com/61284dcff241c2f0729af9f3/61285237ce2e301255d09108_logo-serum.png"
   },
   {
     "symbol": "SOCEAN",
