+++ utils/tokens.tsx
@@ -95,11 +95,7 @@ export async function tryGetMint(
       account,
     }
   } catch (ex) {
-    console.error(
-      `Can't fetch mint ${publicKey?.toBase58()} @ ${connection.rpcEndpoint}`,
-      ex
-    )
-    return undefined
+    console.error(`Can't fetch mint ${publicKey?.toBase58()}`, ex)
   }
 }
